# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The JSON interface to a HyperUplink board, exposing the same\ncategories, forums, topics and administrative surfaces the web\ninterface serves, with every response shaped by the permissions\nof the user the API key belongs to.\n\nAuthentication happens with an API key that a signed-in user\nmints under Account -\u003e API on the board itself, and the secret\nit hands out once is sent either as a bearer token in the\nAuthorization header or verbatim in the X-API-Key header, so a\nrequest that carries neither is answered with 401 before it\never reaches a controller.",
        "title": "HyperUplink API",
        "contact": {
            "name": "HyperUplink",
            "url": "https://hyperup.link"
        },
        "license": {
            "name": "SEGV",
            "url": "https://xn--gckvb8fzb.com/segv/"
        },
        "version": "{{.Version}}"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/comms/xmpp": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Show the XMPP settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "comms_xmpp": {
                                    "$ref": "#/definitions/setting.CommsXMPP"
                                },
                                "target_ids": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update the XMPP settings",
                "parameters": [
                    {
                        "description": "The XMPP settings to store",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/xmpp.UpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/request.ValidationErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
...
        "/admin/permissions/apply": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Grant a group a permission on a category",
                "parameters": [
                    {
                        "description": "The group, the category and the level to grant",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/permissions.ApplyInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/request.ValidationErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
...
        "/admin/permissions/group/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The group identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
...
        "/admin/users": {
            "get": {
                "description": "Banned and deleted users are listed as well, which is what separates\nthis from the public user endpoint.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List the users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "users": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/user.Detail"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The user identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
...
        "/attachments/{attachment}": {
            "get": {
                "description": "The file is streamed back under the name it was uploaded with, and\ncallers who cannot read the category the attachment was posted in\nare answered as though it did not exist.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "board"
                ],
                "summary": "Download an attachment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The attachment identifier",
                        "name": "attachment",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "permissions.ApplyInput": {
            "type": "object",
            "required": [
                "level"
            ],
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string",
                    "maxLength": 32
                },
                "level": {
                    "type": "string",
                    "enum": [
                        "none",
                        "read",
                        "read_write",
                        "read_write_moderate"
                    ]
                }
            }
        },
...
        "pgtype.InfinityModifier": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                0,
                -1
            ],
            "x-enum-varnames": [
                "Infinity",
                "Finite",
                "NegativeInfinity"
            ]
        },
...
        "pgtype.Timestamp": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "description": "Time zone will be ignored when encoding to PostgreSQL.",
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
...
        "request.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "forbidden"
                }
            }
        },
        "request.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "request.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "setting.CommsXMPP": {
            "type": "object",
            "properties": {
                "target_id": {
                    "type": "string"
                }
            }
        },
...
        "user.Detail": {
            "type": "object",
            "properties": {
                "banned_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "confirmed_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "deleted_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "email": {
                    "type": "string"
                },
                "email_is_jid": {
                    "type": "boolean"
                },
                "email_unconfirmed": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "member_of": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "otp_enabled": {
                    "type": "boolean"
                },
                "profile_picture": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/user.Role"
                },
                "sign_in_last_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "signature_html": {
                    "type": "string"
                },
                "signature_text": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "user.Role": {
            "type": "string",
            "enum": [
                "guest",
                "user",
                "admin"
            ],
            "x-enum-varnames": [
                "GuestRole",
                "UserRole",
                "AdminRole"
            ]
        },
...
        "xmpp.UpdateInput": {
            "type": "object",
            "properties": {
                "target_id": {
                    "type": "string",
                    "maxLength": 64
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "description": "The API key secret on its own.",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "The API key secret prefixed with \"Bearer \".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}